Move lru to common.cache package
